dalle/pyproject.toml
[tool.poetry]
name = "dalle"
version = "0.1.0"
description = ""
authors = ["Artem Khokhlov <god.bless.data.science@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.0.1"
pytelegrambotapi = "^4.8.0"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
